requirements.txt
numpy==1.26.4
httpx==0.27.0
pynau7802==0.2.2
mlx90640-driver==1.1.2
mlx90640-driver-devicetree==1.1.0
scipy==1.12.0
cmapy==0.6.6
bme680==1.1.1
pydub==0.25.1
pyaudio==0.2.14
torchaudio==2.2.1
adafruit-circuitpython-neopixel-spi==1.0.9
gpiod==2.1.3